core: allocate opaque data using the system allocator
Since our opaque data may outlive the JS object (and more crucially
JSContext) that they live on, use the system allocator to not depend on
the JS runtime for this.

This is a problem in case we need to uv_close in a finalizer so the
freeing happens later. It's possible the JS runtime was already torn
down.