timers: properly fix timer finalization
We must get rid of all the JS objects in the finalizer, that's rule
number 1.

Now in this case, when clear_timer is called we may drop the single
reference that holds the timer JS object (th->obj) and the finalizer
will run on the spot, de-refing it again, causing an assertion. Avoid it
by using a local variable and reseting its value before freeing it.