core: refactor how JS code is bundled
Instead of relying on running qjsc at build time, use "incbin" to embed
bre-bundled JS, which we bundle with esbuild.

Initially I wanted to keep the build without any dependencies, but this
way simplifies everything so much, it's worth it.

In addition, put the entire stdlib on a single module: @tjs/std.