core: refactor sending / receiving APIs
- read / receive operations will take a Uint8Array where to read into
- write / send operations will take a Uint8Array to send

Technically there is a risk since the array which we reference could be
detached or altered while the operation is in progress, but that's a
compromise we are making. In the future QuickJS may provide a way to
"lock" a typed array while its being worked on on the C side, a la
Python's PyObject_GetBuffer / PyBuffer_Release.