ffi: fixed Pointer.deref() and DlSymbol.addr
* ffi: added tests exercising DlSymbol.addr and Pointer.deref()

This exposes a couple of bugs with these functions. Fixes to come.

* ffi: fixed Pointer.deref() when level > 1

Due to a typo, the type passed to the Pointer constructor during deref()
was always undefined.
Additionally, the Pointer cannot be deref()'ed into another pointer without
actually performing a dereference in C space, otherwise you're treating
the original address as the pointer on every level.

* ffi: fixed typo in DlSymbol.addr causing it to always return undefined