core: fix crash when running a file
Now that the startup logic is in JS, we JS_Eval that file as part of
TJS_Run.

That file, in turn, may evaluate the given file as a module, which
results on abother JS_Eval before we are done with the first.

This seems to corrupt the vm state.

The solution here was to queue a microtask, thus breaking away from the
original problem. A hack to be sure...