core: refactor code bundling for a faster startup
Embed the bytecode directly, as C files. The bytecode is generated with qjsc by running `make js`.

Note: atomics had to be disabled so the bytecode is compatible across platforms. In the future I'll either implement the necessary bits for it to work on Windows, or just stubs that error out.

Fixes: https://github.com/saghul/txiki.js/issues/301