test: fix tests failing on windows
The tests scan stderr of the child process for a particular string. My
hypothesis is that, because they don't wait for the child to terminate,
there is a race window where the parent tries to read stderr before the
child is completely done writing the error message.

This problem doesn't manifest on the other buildbots but maybe the
Windows buildbots are a little slower, or maybe line buffering works
different there, making the problem more apparent.